get rid of binaries
