--- kubemarine/resources/scripts/source/ipip_trns/go.mod
@@ -1,5 +1,0 @@
-module code.google.com/p/gopacket
-
-go 1.19
-
-require github.com/google/gopacket v1.1.19
